edit geojson component
